requirements.txt
pydantic
dotenv
google-genai